nrps_pks: move muscle-using unit tests to integration
